requirements.txt
tgcrypto
asyncio
ffmpeg-python
py-tgcalls
pyrogram
youtube-search-python==1.4.9
yt-dlp==2021.11.10.1
git+https://github.com/ytdl-org/youtube-dl@master
youtube-search
python-dotenv
dnspython
gitpython
aiofiles
aiohttp
requests
pillow
heroku3
motor
psutil
future
wget